moved redux logger configuration to another file.
